+++ src/roa/models/FieldRulePropertySearch.php
@@ -54,9 +54,9 @@ class FieldRulePropertySearch extends FieldRuleProperty implements
 
         return new ActiveDataProvider([
             'query' => $class::find()->andFilterWhere([
-                    'rule_id' => $this->rule_id,
-                    'created_by' => $this->created_by,
-                ]),
+                'rule_id' => $this->rule_id,
+                'created_by' => $this->created_by,
+            ]),
         ]);
     }
 }