+++ src/roa/models/FormSearch.php
@@ -7,6 +7,7 @@ use yii\data\ActiveDataProvider;
 
 /**
  * Contract to filter and sort collections of `Form` records.
+ *
  * @author Angel (Faryshta) Guevara <aguevara@alquimiadigital.mx>
  */
 class FormSearch extends Form implements ResourceSearch