+++ src/roa/models/SectionField.php
@@ -9,9 +9,7 @@ use yii\helpers\Url;
 use yii\web\Link;
 
 /**
- * ROA contract handling form Section records.
- *
- * @method void checkAccess(array $params)
+ * ROA contract handling form SectionField records.
  */
 class SectionField extends base\SectionField implements Contract
 {
@@ -43,21 +41,6 @@ class SectionField extends base\SectionField implements Contract
 
         return array_merge($this->getContractLinks(), [
             'field' => $this->field->getSelfLink(),
-            'curies' => [
-                new Link([
-                    'name' => 'embeddable',
-                    'href' => Url::to($selfLink, ['expand' => '{rel}']),
-                    'title' => 'Embeddable and not Nestable related resources.',
-                ]),
-                new Link([
-                    'name' => 'nestable',
-                    'href' => Url::to($selfLink, ['expand' => '{rel}']),
-                    'title' => 'Embeddable and Nestable related resources.',
-                ]),
-            ],
-            'embeddable:solicitudeValuesDetail' => 'solicitudeValuesDetail',
-            'nestable:field' => 'field',
-            'nestable:section' => 'section',
         ]);
     }