+++ src/roa/models/SolicitudeValue.php
@@ -9,9 +9,7 @@ use yii\helpers\Url;
 use yii\web\Link;
 
 /**
- * ROA contract handling Field records.
- *
- * @method void checkAccess(array $params)
+ * ROA contract handling SolicitudeValue records.
  */
 class SolicitudeValue extends base\SolicitudeValue implements Contract
 {
@@ -49,18 +47,6 @@ class SolicitudeValue extends base\SolicitudeValue implements Contract
         return array_merge($this->getContractLinks(), [
             'field' => $this->field->getSelfLink(),
             'section' => $this->section->getSelfLink(),
-            'sectionField' => $this->sectionField->getSelfLink(),
-            'curies' => [
-                new Link([
-                    'name' => 'nestable',
-                    'href' => Url::to($selfLink, ['expand' => '{rel}']),
-                    'title' => 'Embeddable and Nestable related resources.',
-                ]),
-            ],
-            'nestable:field' => 'field',
-            'nestable:section' => 'section',
-            'nestable:sectionField' => 'sectionField',
-            'nestable:solicitude' => 'solicitude',
         ]);
     }